win32/requirements.txt
opencv-python==4.7.0.72
pywebview==4.1
